# startUp/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AiEd Learning Assistant",
    "version": "1.0.0",
    "description": "AiEd Learning Assistant API Documentation"
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "quarters": {
        "type": "object",
        "required": [
          "title",
          "description",
          "year",
          "human_date",
          "start_date",
          "end_date",
          "index",
          "id"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "year": {
            "type": "string"
          },
          "human_date": {
            "type": "string"
          },
          "start_date": {
            "type": "string"
          },
          "end_date": {
            "type": "string"
          },
          "index": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "cover": {
            "type": "string",
            "nullable": true
          },
          "color_primary": {
            "type": "string",
            "nullable": true
          },
          "color_primary_dark": {
            "type": "string",
            "nullable": true
          },
          "introduction": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/{lang}/quarters": {
      "post": {
        "summary": "Create Quarter",
        "description": "Create Quarter",
        "tags": ["Quarters"],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/quarters"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "summary": "Get all Quarters",
        "description": "Get all Quarters by language",
        "tags": ["Quarters"],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "en"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/quarters"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/{lang}/quarters/{quarter_id}": {
      "get": {
        "summary": "Get Specific Quarter",
        "description": "Get Specific Quarter",
        "tags": ["Quarters"],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "en"
          },
          {
            "name": "quarter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2023_03"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/quarters"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Quarter",
        "description": "Update Quarter",
        "tags": ["Quarters"],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quarter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/quarters"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/quarters"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Quarter",
        "description": "Delete Quarter",
        "tags": ["Quarters"],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quarter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/v1/{lang}/quarters/{quarter_id}/lessons/{lesson_id}/days/{day_id}/read": {
      "get": {
        "summary": "Array of Days Read Content",
        "description": "Array of Days Read Content",
        "tags": ["Days"],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "en"
          },
          {
            "name": "quarter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2023_03"
          },
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "01"
          },
          {
            "name": "day_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "01"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}